requirements.txt
appdirs==1.4.4
astor==0.8.1
attrdict==2.0.1
Babel==2.11.0
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
black==21.7b0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
cython-bbox==0.1.3
decorator==5.1.1
dill==0.3.6
et-xmlfile==1.1.0
fire==0.5.0
Flask==2.2.2
Flask-Babel==2.0.0
fonttools==4.38.0
future==0.18.2
idna==3.4
imageio==2.23.0
imgaug==0.4.0
isort==5.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lmdb==1.4.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
multiprocess==0.70.14
mypy-extensions==0.4.3
networkx==2.8.8
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==22.0
paddle-bfloat==0.1.7
paddleocr==2.6.1.1
pandas==1.5.2
pathspec==0.10.3
pdf2docx==0.5.6
Pillow==9.3.0
premailer==3.10.0
protobuf==3.20.0
pyclipper==1.3.0.post4
pycryptodome==3.16.0
PyMuPDF==1.20.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==2.13.6
regex==2022.10.31
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
shapely==2.0.0
six==1.16.0
soupsieve==2.3.2.post1
termcolor==2.1.1
tifffile==2022.10.10
tomli==1.2.3
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
visualdl==2.4.1
Werkzeug==2.2.2

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1  # Google Colab version
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
protobuf<4.21.5  # https://github.com/ultralytics/yolov5/issues/8012

pandas>=1.1.4
seaborn>=0.11.0

ipython  
psutil  
thop  
streamlit
wget

# lidar 2 camera
open3d==0.16.0
pypotree

# lidar view
moviepy
